requirements.txt
httpx>=0.27
playwright>=1.44
pyyaml>=6.0
anthropic>=0.28
